optimization and some fixes

+++ basic/utils/stack.h
@@ -1,9 +1,12 @@
 #ifndef LOADER_STACK_H
 #define LOADER_STACK_H
 
+#include <cstdlib>
+
+// stack can grow upwards or downwards
 struct stack {
     void* ptr;
-    unsigned long long size;
+    size_t size;
 };
 
 #endif //LOADER_STACK_H
